Day 6 Part 2
Man that parsing was annoying, and the test input doesn't match the
actual input (various spacing for the last sum), which means that
I needed to alter the input to get this to work, by adding a number of
spaces after the operator that represents the number of items in the
calculation. Rather annoying!